--- a/src/haywire/http_request.c
+++ b/src/haywire/http_request.c
@@ -102,7 +102,7 @@ http_request* create_http_request(http_connection* connection)
     request->body = malloc(sizeof(hw_string));
     request->body->value = NULL;
     request->body->length = 0;
-    request->start_t = clock();
+    request->begin_time = clock();
     INCREMENT_STAT(stat_requests_created_total);
     return request;
 }