Adding timing support. It will allow us to check how long each request takes to process.