@kellabyte 2016-01-27
Yeah, I was concerned getting wallclock every request or 2x per request would be expensive. That was my motivation to writing the HTTP partial response cache to avoid having to call the wallclock every request to get the date/time header value.