@nmdguerreiro 2016-01-28
This can be replaced by #96