@kellabyte 2016-01-29
Closing since #96 replaces this.